testing-docker/requirements.txt
Flask==2.0.1
mysql-connector-python==8.0.25
Werkzeug==2.0.1
pyjwt==2.1.0
flask-cors==3.0.10